+++ kanban.server/server/services/ListsService.js
@@ -10,8 +10,8 @@ class ListsService {
     return await dbContext.Lists.create(body)
   }
 
-  async deleteList(id) {
-    const data = await dbContext.Lists.findOneAndDelete({ _id: id })
+  async deleteList(id, userId) {
+    const data = await dbContext.Lists.findOneAndDelete({ _id: id, creatorId: userId })
     if (!data) {
       throw new BadRequest('Invalid Id - Delete')
     }
